Update readme and routes

+++ routes/admin.php
@@ -6,17 +6,9 @@ Route::group([
     'namespace'  => 'Admin',
     'middleware' => ['web', 'laradium'],
 ], function () {
-    Route::delete('content-block/{id}', [
-        'uses' => '\Laradium\Laradium\Content\Http\Controllers\Admin\PageController@contentBlockDelete'
-    ]);
+    Route::delete('content-block/{id}', '\Laradium\Laradium\Content\Http\Controllers\Admin\PageController@contentBlockDelete');
 
-    Route::get('pages/create/{channel}', [
-        'uses' => '\Laradium\Laradium\Content\Base\Resources\PageResource@create',
-        'as'   => 'pages.create'
-    ]);
+    Route::get('pages/create/{channel}', '\Laradium\Laradium\Content\Base\Resources\PageResource@create')->name('pages.create');
 
-    Route::get('pages/{page}/edit', [
-        'uses' => '\Laradium\Laradium\Content\Base\Resources\PageResource@edit',
-        'as'   => 'pages.edit'
-    ]);
+    Route::get('pages/{page}/edit', '\Laradium\Laradium\Content\Base\Resources\PageResource@edit')->name('pages.edit');
 });
